Separate DOM functions out of newBookings
